% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map_df}
\alias{map_df}
\title{GIS data for boundaries at Region, STP/ICS and Local Authority level}
\format{
A \code{sf} object, with 4 fields
}
\source{
{
  Data and Advanced Analytics Team at NHS BSA, using data from the API 
  provided by \href{https://www.arcgis.com/index.html}{ArcGIS}
}
}
\usage{
data(map_df)
}
\description{
\itemize{
  \item GEOGRAPHY. one of Region, STP/ICS or Local Authority
  \item SUB_GEOGRAPHY_CODE. region or organisation code
  \item SUB_GEOGRAPHY_NAME. region or organisation name
  \item GEOMETRY. boundary polygons
}
}
\keyword{datasets}
