% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metrics_by_breakdown_and_ch_flag_df}
\alias{metrics_by_breakdown_and_ch_flag_df}
\title{Aggregated prescribing data, with breakdowns by geography, demography,
gender, age band and carehome type}
\format{
A data frame with 784 rows and 22 variables
}
\source{
{
  Data and Advanced Analytics Team at NHS BSA, using 
  \href{https://www.cqc.org.uk/about-us/transparency/using-cqc-data}{CQC} and 
  \href{https://beta.ordnancesurvey.co.uk/products/addressbase-plus}{AddressBase+} 
  data in combination with internal data
}
}
\usage{
data(metrics_by_breakdown_and_ch_flag_df)
}
\description{
\itemize{
  \item BREAKDOWN. overall breakdown category (Overall, Geographical - Region,
    Geographical - STP/ICS, Geographical - Local Authority, Demographical - Gender,
    Demographical - Age Band, Additional - Gender and Age Band,
    Additional - Care home type)
  \item SUB_BREAKDOWN_CODE. region or organisational codes (geography only)
  \item SUB_BREAKDOWN_NAME. name in breakdown (not in additional)
  \item GENDER. Female or Male (Additional - Gender and Age Band only)
  \item AGE_BAND. 65-69, ..., 85-89, 90+ (Additional - Gender and Age Band only)
  \item NURSING_HOME_FLAG. 1 if a nursing home, 0 otherwise (Additional - Care home type only)
  \item RESIDENTIAL_HOME_FLAG. 1 if a residential home, 0 otherwise (Additional - Care home type only)
  \item TOTAL_PATIENTS.number of distinct patients
  \item ITEMS_PER_PATIENT_MONTH. average number of distinct items per patient per month
  \item COST_PER_PATIENT_MONTH. average cost of items per patient per month
  \item TOTAL_PATIENTS_UNIQUE_MEDICINES. total unique medicines
  \item UNIQUE_MEDICINES_PER_PATIENT_MONTH average number of unique medicines per patient per month
  \item TOTAL_PATIENTS_TEN_OR_MORE total number of patients prescribed 10 or more unique medicines
  \item PCT_PATIENTS_TEN_OR_MORE_PER_PATIENT_MONTH percentage of patients
    prescribed 10 or more unique medicines per month
  \item SDC_*. Same 7 aggregate patient data fields, but with Statistical
    Disclosure Control and rounding
}
}
\keyword{datasets}
