% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metrics_by_geo_and_ch_flag}
\alias{metrics_by_geo_and_ch_flag}
\title{Aggregated prescribing data, with breakdowns by geography, demography,
gender, age band and carehome type}
\format{
A data frame with 784 rows and 22 variables
}
\source{
{ Data and Advanced Analytics Team at NHS BSA, using
  \href{https://www.cqc.org.uk/about-us/transparency/using-cqc-data}{CQC} and
  \href{https://beta.ordnancesurvey.co.uk/products/addressbase-plus}{AddressBase+}
  data in combination with internal data }
}
\usage{
data(metrics_by_geo_and_ch_flag)
}
\description{
\itemize{
  \item FY. financial year, one of 2020/21, 2021/22 or 2022/23
  \item GEOGRAPHY. Region, ICB or Local Authority
  \item SUB_GEOGRAPHY_CODE. region or organisational codes
  \item SUB_GEOGRAPHY_NAME. name in breakdown
  \item CH_FLAG. values apply to Carehomes, \code{TRUE} or \code{FALSE}
  \item TOTAL_PATIENTS.number of distinct patients
  \item ITEMS_PPM. average number of distinct items per patient per month
  \item COST_PPM. average cost of items per patient per month
  \item TOTAL_PATIENTS_UNIQ_MED. total unique medicines
  \item UNIQ_MEDS_PPM. average number of unique medicines per patient per 
    month
  \item TOTAL_PATIENTS_GTE_TEN total number of patients prescribed 10 or more
    unique medicines
  \item PCT_PX_GTE_TEN_PPM percentage of patients prescribed 10 or more 
    unique medicines per month
}
}
\keyword{datasets}
