% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo_data}
\alias{geo_data}
\title{GIS data for boundaries at Region, ICS and Local Authority level}
\format{
A named \code{list}, created from \href{https://geojson.org/}{geojson}
  using \code{jsonlite::fromJSON(simplifyVector = FALSE)}
}
\source{
{ Data and Advanced Analytics Team at NHS BSA, using data from the
  API provided by \href{https://www.arcgis.com/index.html}{ArcGIS} }
}
\usage{
data(geo_data)
}
\description{
\itemize{
    \item Region. \code{Featurecollection} of 7 \code{Feature}s
    \item ICS. \code{Featurecollection} of 42 \code{Feature}s
    \item Local Authority. \code{Featurecollection} of 309 \code{Feature}s
}
}
\keyword{datasets}
