% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metrics_by_ch_type}
\alias{metrics_by_ch_type}
\alias{metrics_by_ch_type_df}
\title{Aggregated prescribing data broken down by financial year and carehome type}
\format{
A data frame with 12 rows and 14 variables
}
\source{
{ Data and Advanced Analytics Team at NHS BSA, using
  \href{https://www.cqc.org.uk/about-us/transparency/using-cqc-data}{CQC} and
  \href{https://beta.ordnancesurvey.co.uk/products/addressbase-plus}{AddressBase+}
  data in combination with internal data }
}
\usage{
data(metrics_by_ch_type_df)
}
\description{
\itemize{
  \item FY. financial year, one of 2020/21, 2021/22 or 2022/23
  \item CH_TYPE. one of Carehome, Non-carehome, Nursing Home, Residential Home
  \item ITEMS_PPM. Mean number of items prescribed per patient month
  \item COST_PPM. Mean cost of items prescribed per patient month
  \item UNIQ_MEDS_PPM. Mean number of unique medicines from BNF Chapters 1-4,
    6-10 prescribed per patient month
  \item UNIQ_MEDS_FALLS_PPM. Mean number of unique medicines associated with
    falls risk prescribed per patient month
  \item TOTAL_PM. Total patient months
  \item TOTAL_PM_ACB. total number of patient months in which at least 2
    medicines of moderate to high anticholinergic burden were prescribed
  \item TOTAL_PM_DAMN. total number of patient months in which at least 2
    medicines likely to cause kidney injury were prescribed
  \item PCT_PM_GTE_SIX. Percentage of patient months in which at least 6 
    unique medicines from BNF Chapters 1-4, 6-10 were prescribed
  \item PCT_PM_GTE_TEN. Percentage of patient months in which at least 10
    unique medicines from BNF Chapters 1-4, 6-10 were prescribed
  \item PCT_PM_ACB. Percentage of patient months with anticholinergic 
    prescribing, in which at least 2 medicines of moderate to high 
    anticholinergic burden were prescribed
  \item PCT_PM_DAMN. Percentage of patient months with DAMN prescribing, in
    which at least 2 medicines likely to cause kidney injury were prescribed
  \item PCT_PM_FALLS. Percentage of patient months in which at least 3 unique
   medicines associated with falls risk were prescribed
}
}
\keyword{datasets}
